fix album like

+++ src/api/album-likes/handler.js
@@ -1,3 +1,4 @@
+const ClientError = require('../../exceptions/ClientError')
 const { successResponse } = require('../../utils/response')
 const autoBind = require('auto-bind')
 
@@ -9,17 +10,20 @@ class AlbumLikesHandler {
     autoBind(this)
   }
 
-  async countHandler (request) {
+  async countHandler (request, h) {
     const { id } = request.params
 
     await this._albumService.getById(id)
 
-    const likes = await this._service.countLikes(id)
-    return successResponse({
-      datas: {
+    const { source, count: likes } = await this._service.countLikes(id)
+    const response = h.response(successResponse({
+      data: {
         likes
       }
-    })
+    }))
+    if (source === 'cache') response.header('X-Data-Source', 'cache')
+
+    return response
   }
 
   async postHandler (request, h) {
@@ -27,6 +31,12 @@ class AlbumLikesHandler {
     const { id: userId } = request.auth.credentials
 
     await this._albumService.getById(albumId)
+    const isLiked = await this._service.checkLike({
+      albumId,
+      userId
+    })
+
+    if (isLiked) throw new ClientError('Album sudah pernah disukai')
 
     await this._service.addLike({
       albumId,
@@ -45,6 +55,12 @@ class AlbumLikesHandler {
     const { id: userId } = request.auth.credentials
 
     await this._albumService.getById(albumId)
+    const isLiked = await this._service.checkLike({
+      albumId,
+      userId
+    })
+
+    if (!isLiked) throw new ClientError('Album belum pernah disukai')
 
     await this._service.removeLike({
       albumId,
